move toggle code to a separate file
