+++ plugin/inlay_hint.py
@@ -1,44 +1,13 @@
-from .core.settings import userprefs
+from .inlay_hints_toggle import InlayHints, ToggleInlayHintStrategy
 from .core.protocol import InlayHintLabelPart, MarkupContent, Point, InlayHint, Request
 from .core.registry import LspTextCommand,  windows
 from .core.sessions import Session
-from .core.typing import Optional, Union, Literal
+from .core.typing import Optional, Union
 from .core.views import point_to_offset
 from .formatting import apply_text_edits_to_view
 import html
 import sublime
 import uuid
-
-ToggleInlayHintStrategy = Literal["current_view", "current_window", "all_windows"]
-
-
-class InlayHints:
-    toggle_strategy = "current_view"  # type: ToggleInlayHintStrategy
-    global_show_inlay_hints = False
-
-    @staticmethod
-    def get_target(v: sublime.View) -> Union[sublime.View, sublime.Window]:
-        if InlayHints.toggle_strategy == 'current_window':
-            w = v.window()
-            if not w:
-                raise Exception('no window')
-            return w
-        return v
-
-    @staticmethod
-    def are_enabled(v: sublime.View) -> bool:
-        if InlayHints.toggle_strategy == 'all_windows':
-            return InlayHints.global_show_inlay_hints
-        target = InlayHints.get_target(v)
-        return target.settings().get('lsp_show_inlay_hints') or userprefs().show_inlay_hints
-
-    @staticmethod
-    def toggle(v: sublime.View) -> None:
-        if InlayHints.toggle_strategy == 'all_windows':
-            InlayHints.global_show_inlay_hints = not InlayHints.global_show_inlay_hints
-            return
-        target = InlayHints.get_target(v)
-        target.settings().set('lsp_show_inlay_hints', not InlayHints.are_enabled(v))
 
 
 class LspToggleInlayHintsCommand(LspTextCommand):
