+++ plugin/session_buffer.py
@@ -1,3 +1,4 @@
+from .inlay_hints_toggle import InlayHints
 from .core.protocol import Diagnostic
 from .core.protocol import DiagnosticSeverity
 from .core.protocol import DocumentLink
@@ -31,7 +32,7 @@ from .core.views import range_to_region
 from .core.views import region_to_range
 from .core.views import text_document_identifier
 from .core.views import will_save
-from .inlay_hint import InlayHints, inlay_hint_to_phantom
+from .inlay_hint import inlay_hint_to_phantom
 from .semantic_highlighting import SemanticToken
 from functools import partial
 from weakref import WeakSet
